feat: add cancel-subscription ivy interface
Also changed how the `defvar`s are working based on [this
post][defvar-post]. They are all declared as `nil` initially and then
set immediately after their declaration via `setq`.

Updated the `ivy-feedwrangler--base-url` to not include the
"feed_items" portion of the path so it can be used in more places.

Added `json-parse!` macro which cleans up how JSON parsing from a
`url-retrieve` works.

[defvar-post]: http://ergoemacs.org/emacs/elisp_defvar_problem.html